move snapshot test to a different file
