+++ src/components/Home.js
@@ -6,7 +6,7 @@ const Home = () => (
     <Nav />
     <h1>Welcome to our page!</h1>
     <p>
-      Lorem ipsum dolor sit amet, consectetur adipisicing elit,
+      This is my calculator app, developed using react
       sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
       enim ad minim veniam, quis nostrud exercitation ullamco laboris
       nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in
